refactor(audience): extract AudienceAjaxController from AudienceLoader
Item 3 / sprint 16 — final fragmentation, completing the item. AudienceLoader
goes 1131 → 408 lines.

The ~13 admin-ajax handlers (bookings, conflicts, schedule slots, user
search, per-user permissions, custom-fields editor) plus their two private
helpers (sanitize_dependent_groups, preserve_dependent_labels) move verbatim
into a new final class AudienceAjaxController, which `use`s the same
AjaxTrait and owns a register() that wires the 13 wp_ajax_* hooks. The loader
now keeps only bootstrap/hooks/enqueue and delegates:
$this->ajax_controller = new AudienceAjaxController(); ->register(). The
loader drops `use AjaxTrait` (no longer used outside the handlers).

Pure code movement, PHPStan-verified (every symbol resolves). Test-first: the
three helper tests that previously reflected into AudienceLoader now target
AudienceAjaxController, joined by a hook-registration test and two handler
characterizations (create-booking, check-conflicts). AudienceLoaderTest stays
green at 11/11 — its hook-registration test passes because the controller
registers the same action names.

Verified locally: full PHPUnit 4895 tests / 12573 assertions pass, PHPStan
level 8 clean, WPCS clean, php -l clean. Item 3 complete (candidate-edit s13,
rate-limit s14, recruitment-admin-page s15, audience-loader s16).
